+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@ Release Notes
 =============
 ## 1.6.15
 * Key Vaults: Allow deploying standalone secrets without a KeyVault in the same deployment
+* WebApp/Functions: no longer overwrites production slot settings when using a multi-slot deploy
 
 ## 1.6.14
 * Container Service (AKS): Adds `kubelet_identity` operator to suppor a user assigned identity for kubelet.